+++ Engine/box.h
@@ -6,18 +6,18 @@
 class box
 {
 private:
-	int x;
-	int y;
-	int dimension = 20;
+	float x;
+	float y;
+	float dimension = 20.0f;
 	bool empty = true;
 	Color c = { 127,0,0 };
 public:
 	void composebox(Graphics &gfx)
 	{
-		gfx.DrawRectDim(x, y, dimension, dimension, c);
+		gfx.DrawRectDim(int(x), int(y), int(dimension), int(dimension), c);
 	}
 	void changecolor();
 	bool collision(Face& face);
-	box(int x_0, int y_0);
-	void changexy(int x_0, int y_0);
+	box(float x_0, float y_0);
+	void changexy(float x_0, float y_0);
 };
